+++ packages/commerce-sdk-react/src/hooks/ShopperLogin/index.ts
@@ -4,4 +4,4 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-export * from './mutation'
+export * from './helper'
